+++ src/records/Address.php
@@ -14,7 +14,7 @@ use craft\db\Table;
  * Class Address record.
  *
  * @property int $id ID
- * @property int $ownerId Owner ID
+ * @property int|null $ownerId Owner ID
  * @property string $countryCode Country code
  * @property string|null $administrativeArea Administrative area
  * @property string|null $locality Locality